Added ProjectMap class.
Added a bunch of getters (for example User.getAllForUser(Long userId)) that are designed to return a collection of objects that is filtered to only include items a particular user is supposed to have access to. 

Admin screens are now only visible to admins.

Several of the collections that are loaded by default in the Controller (for example zones, projects, users, etc...) will now filter results based on what the user should have access to. This should limit the choices available in the drop downs on the issue search screen.