Hugely speed up rendering of sortableCell.tag from ~10ms to ~0ms.
The speedup appears to be from explicitly declaring all tag attributes (like style and cssClass) when calling the tag. As well as setting an empty 'sortIcon' instead of leaving it undefined. Best guess as to why, is that explicitly declaring everything prevents the jsp engine from looking all over the context for those values.

On dashboard.jsp, which can easily have 6 issue forms each containing 5 sortableCells, this saves about 10*6*5=300ms of jsp rendering time for the page.